help explain split part 1
