+++ src/ai/help.rs
