+++ tests/api/ai_explain.rs
@@ -7,7 +7,8 @@ use diesel::{QueryDsl, RunQueryDsl};
 use hmac::Mac;
 use rumba::ai::constants::AI_EXPLAIN_VERSION;
 use rumba::ai::explain::{hash_highlighted, ExplainRequest, HmacSha256};
-use rumba::db::ai::{add_explain_answer, ExplainFeedback, FeedbackTyp};
+use rumba::db::ai_explain::{add_explain_answer, ExplainFeedback};
+use rumba::db::ai_help::FeedbackTyp;
 use rumba::db::model::{AIExplainCacheInsert, AIExplainCacheQuery};
 use rumba::db::schema::ai_explain_cache;
 use rumba::settings::SETTINGS;
